fix: join class and DAO in single file
