+++ include/car_park/car_park.hpp
@@ -1,27 +1,9 @@
-/**
- * @brief One line description of the car_park library
- *
- * More information about the car_park library
- * can go here.
- *
- * @see https://github.com/fpmi-tp2024/tpmp-gpd-lab5-death-pact
- */
+
 #pragma once
 
 namespace car_park {
 
-/**
- * @brief One line description of add_one goes here
- *
- * For a more complicated function, here
- * would be more information about what it does.
- *
- * @param x Description of the input parameter x
- *
- * @note Something particular to note about add_one
- *
- * @warning A warning about add_one
- */
+
 int add_one(int x);
 
 }
